requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
py-clob-client>=0.34.0
web3>=7.0.0
websocket-client>=1.6.0
